Renamed passage to jodit-html
