+++ examples/data/joditHtml.json
@@ -1,5 +1,5 @@
 {
-  "type": "PASSAGE",
+  "type": "JODIT_HTML",
   "position": 1,
   "data": {
       "width": 12,
