+++ src/teaching-element/index.vue
@@ -31,7 +31,7 @@ import TeEmbed from './Embed.vue';
 import TeHtml from './Html.vue';
 import TeImage from './Image.vue';
 import TeModal from './Modal/index.vue';
-import TePassage from './Passage.vue';
+import TeJoditHtml from './JoditHtml.vue';
 import TePdf from './Pdf.vue';
 import TeTable from './Table.vue';
 import TeVideo from './Video.vue';
@@ -66,7 +66,7 @@ export default {
     TeHtml,
     TeImage,
     TeModal,
-    TePassage,
+    TeJoditHtml,
     TePdf,
     TeTable,
     TeVideo
